refactor: NotificationManager 클래스 생성하여 Notification에 관한 책임 분리
